+++ src/external-types/emoji-mart/dist-es/utils/emoji-index/nimble-emoji-index.d.ts
@@ -1,0 +1,15 @@
+import 'emoji-mart'
+
+declare module 'emoji-mart' {
+  export interface SearchOption {
+    emojisToShowFilter: (emoji: EmojiData) => boolean
+    maxResults: number,
+    include: EmojiData[]
+    exclude: EmojiData[]
+    custom: EmojiData[]
+  }
+
+  export class NimbleEmojiIndex {
+    search (query: string, options: SearchOption): EmojiData[] | null;
+  }
+}